🔨 update landcover loading calls after merging
Just updating the loading calls. Most of the tests in this script now return the same values, as I fixed the loading calls properly. However, the dimensions difference bug may still happen. There also still seems to be a difference in the landcover values when changing the extent directly when coarsening, but the current solution seems fine (and I can't even investigate it for all landcover variables as some full resolutions files have a bug with GDAL)