🗃 archive less important scripts
I'm archiving all the scripts from others/ that are not important for some side task (e.g. data_glossary.jl or data_export-values.jl) or related to BARTs (even if not related to the main BART workflow).

Most of these have useful functions or things which I might reuse, so the archiving makes sense as a 'keep for later use'. Hopefully it'll also be easier to distinguish which scripts to maintain or not.